# contracts/rest-openapi.v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PositionSizingRequest": {
        "description": "Request for sizing a position.",
        "properties": {
          "available_capital": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Capital"
          },
          "lot_size": {
            "default": 1,
            "title": "Lot Size",
            "type": "integer"
          },
          "max_position_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Value"
          },
          "max_position_weight_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Weight Pct"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "risk_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Budget"
          },
          "stop_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Pct"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "target_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Value"
          },
          "target_weight_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Weight Pct"
          }
        },
        "required": [
          "symbol",
          "method"
        ],
        "title": "PositionSizingRequest",
        "type": "object"
      },
      "RebalanceRequest": {
        "description": "Request for a dry-run rebalance plan.",
        "properties": {
          "cash_buffer_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0",
            "title": "Cash Buffer Pct"
          },
          "drift_threshold_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "1",
            "title": "Drift Threshold Pct"
          },
          "liquidate_unmentioned": {
            "default": false,
            "title": "Liquidate Unmentioned",
            "type": "boolean"
          },
          "lot_size": {
            "default": 1,
            "title": "Lot Size",
            "type": "integer"
          },
          "max_position_weight_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Weight Pct"
          },
          "target_weights": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ]
            },
            "title": "Target Weights",
            "type": "object"
          }
        },
        "required": [
          "target_weights"
        ],
        "title": "RebalanceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Kodiak headless trading and research API. All endpoints live under /v1/. Authenticate with: Authorization: Bearer <KODIAK_API_TOKEN>.",
    "title": "Kodiak REST API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/portfolio/analytics": {
      "get": {
        "description": "Get portfolio analytics with transaction-level reconstruction when available.",
        "operationId": "analytics_v1_portfolio_analytics_get",
        "parameters": [
          {
            "in": "query",
            "name": "lookback_days",
            "required": false,
            "schema": {
              "default": 252,
              "minimum": 1,
              "title": "Lookback Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "benchmark_symbol",
            "required": false,
            "schema": {
              "default": "SPY",
              "title": "Benchmark Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analytics V1 Portfolio Analytics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics",
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/v1/portfolio/movers": {
      "get": {
        "description": "Get top movers.",
        "operationId": "movers_v1_portfolio_movers_get",
        "parameters": [
          {
            "in": "query",
            "name": "market_type",
            "required": false,
            "schema": {
              "default": "stocks",
              "title": "Market Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Movers V1 Portfolio Movers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Movers",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/v1/portfolio/position-size": {
      "post": {
        "description": "Calculate a target size for a symbol.",
        "operationId": "position_size_v1_portfolio_position_size_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionSizingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Position Size V1 Portfolio Position Size Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Position Size",
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/v1/portfolio/quote/{symbol}": {
      "get": {
        "description": "Get a quote for a symbol.",
        "operationId": "quote_v1_portfolio_quote__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Quote V1 Portfolio Quote  Symbol  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/v1/portfolio/rebalance-plan": {
      "post": {
        "description": "Generate a dry-run rebalance plan.",
        "operationId": "rebalance_plan_v1_portfolio_rebalance_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rebalance Plan V1 Portfolio Rebalance Plan Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebalance Plan",
        "tags": [
          "portfolio"
        ]
      }
    },
...
  }
}